making category ids ints, converting all others to uuid

+++ HSAPI/src/main/kotlin/com/hobbystock/entities/CategoryEntity.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 @Entity
 @Table(name = "categories")
 data class CategoryEntity(
-        @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val id: Long = 0,
+        @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val id: Int = 0,
         @Column(name = "name", nullable = false, unique = true) val name: String,
         @Column(name = "description") val description: String? = null,
         @Column(name = "created_at") val createdAt: LocalDateTime = LocalDateTime.now(),
